integration-tests-2/requirements.txt
confluent_kafka==2.4.0
docker-compose>=1.20.0
docker==6.1.3
ess-streaming-data-types
pytest
pytest-rerunfailures